Add permission restrictions for feature flags
* add can_update and protected to feature flag entity -
needed to show protected badge and disabled toggles

* create feature flag scopes for a protected environment when we
  create or update it

* Check if user has access to environment on update of feature flag

* Check permissions on feature flag deletion

* Fix new form for feature flags with permissions

+++ ee/app/services/feature_flags/destroy_service.rb
@@ -3,6 +3,9 @@
 module FeatureFlags
   class DestroyService < FeatureFlags::BaseService
     def execute(feature_flag)
+      permission_error = check_permissions(feature_flag)
+      return error(permission_error) if permission_error
+
       ActiveRecord::Base.transaction do
         if feature_flag.destroy
           save_audit_event(audit_event(feature_flag))
@@ -18,6 +21,18 @@ module FeatureFlags
 
     def audit_message(feature_flag)
       "Deleted feature flag <strong>#{feature_flag.name}</strong>."
+    end
+
+    def check_permissions(feature_flag)
+      return unless permissions_enabled?
+
+      feature_flag.scopes.each do |scope|
+        unless environment_accessible_to_user?(scope.environment_scope)
+          return permission_error_message(scope.environment_scope)
+        end
+      end
+
+      nil
     end
   end
 end
